requirements.txt
# Skills-Fabrik Patterns Plugin Dependencies
# Python 3.10+

# Core dependencies
pyyaml>=6.0
psutil>=5.9.0
returns>=0.22.0  # Functional programming: Result, Maybe, Either types
ruff>=0.8.0  # Python formatter + linter (replaces black, flake8, isort)

# Testing (dev only)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Type checking (dev only)
mypy>=1.5.0
